svn merge -r 27229:27471 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@27480 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/include/dart_native_api.h
@@ -53,17 +53,17 @@ typedef struct _Dart_CObject {
     char* as_string;
     char* as_bigint;
     struct {
-      int length;
+      intptr_t length;
       struct _Dart_CObject** values;
     } as_array;
     struct {
       Dart_TypedData_Type type;
-      int length;
+      intptr_t length;
       uint8_t* values;
     } as_typed_data;
     struct {
       Dart_TypedData_Type type;
-      int length;
+      intptr_t length;
       uint8_t* data;
       void* peer;
       Dart_WeakPersistentHandleFinalizer callback;
